==> notebooks/1 - TimeTable Wrangling.ipynb <==
# Timetable data wrangling

In this pyspark notebook, we will wrangle timetable data to process it into the desired data format.

Indeed, we want to modelize our data through a graph, where each node will represent a station with the 15 km boundaries of Zurich HB. The edges will represent any trips from one stop to another.

Thus, we want to convert the timetable data to this graph representation. The end goal of this pyspark kernel is to have 3 tables:
* *Graph nodes*: that will represent the node of our graph (stop_id, stop_name, stop_lat, stop_lon)
* *Graph edges 1*, that will represent one possible set of edges for the graph. This one will include all possible trips from one stop to anothers. Thus, it will includes the public transportation trips, but also the walking trips for all stops. On this graph edges, we will have only one row for a walking trip, without putting any departure time and arrival time.
* *Graph Edges 2*, that will represent another possible set of edges for the graph. The public transportation trips which are the same as *Graph Edges 1*. The key difference is that for any arrival at a stop with a public transportation, we will add all of the possible walking trips from this stop. The departure time of these walking trips will be the same as the arrival time to the stop.

We will then write in hdfs the wrangled datasets

## Creating spark application and libraries import

%%local
import os
import json
from IPython import get_ipython

username = os.environ['JUPYTERHUB_USER']
namespace = os.environ['CI_NAMESPACE']
project = os.environ['CI_PROJECT']

configuration = dict(
    name = f"{username}-namespace-project",
    executorMemory = "4G",
    executorCores = 4,
    numExecutors = 10,
    conf = {
        # "spark.pyspark.python": "/opt/anaconda3/bin/python3", # Use python3
        "spark.jars.repositories": "https://repos.spark-packages.org",
        "spark.jars.packages": "graphframes:graphframes:0.7.0-spark2.3-s_2.11"
    }
)


get_ipython().run_cell_magic('configure', line="-f", 
                             cell=json.dumps(configuration))

spark

sc.addPyFile('graphframes_graphframes-0.7.0-spark2.3-s_2.11.jar')

import pandas
import datetime
import pyspark.sql.functions as F

from graphframes import *
from math import radians, cos, sin, asin, sqrt, floor
from pyspark.sql import Window
from pyspark.sql.types import StringType

## Data loading and exploration

istdaten =  spark.read.orc("/data/sbb/orc/istdaten")
stop = spark.read.orc("/data/sbb/orc/geostops")
calendar = spark.read.csv("/data/sbb/csv/timetable/calendar/2019/05/07/calendar.csv", header=True)
route = spark.read.csv("/data/sbb/csv/timetable/routes/2019/05/07/routes.csv", header=True)
stop_time = spark.read.csv("/data/sbb/csv/timetable/stop_times/2019/05/07/stop_times.csv", header=True)
trip = spark.read.csv("/data/sbb/csv/timetable/trips/2019/05/07/trips.csv", header=True)

istdaten.show(5)

stop.show(5)

calendar.show(5)

route.show(5)

stop_time.show(5)

trip.show(5)

print("istdaten has " + str(istdaten.count())+ " rows")
print("stops has " + str(stop.count()) + " rows")
print("calendars has " + str(calendar.count()) +" rows")
print("routes has "+ str(route.count()) + " rows")
print("stop_times has " + str(stop_time.count()) + " rows")
print("trips has " + str(trip.count()) +" rows")

## Filtering

We want to keep only stops within a 15 km radius from Zürich HBB. 

We also want to keep only trips on business hours (8-20) and planned every business days

def haversine_distance(lon1, lat1, lon2, lat2):
    """
    Calculate the great circle distance between two points 
    on the earth (specified in decimal degrees)
    """
    # convert decimal degrees to radians 
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    # haversine formula 
    dlon = lon2 - lon1 
    dlat = lat2 - lat1 
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * asin(sqrt(a)) 
    r = 6371 # Radius of earth in kilometers. Use 3956 for miles
    return c * r

lat_zurich = 47.378177
lon_zurich = 8.540192

@F.udf
def in_zurich(lat, lon):
    return haversine_distance(lon, lat, lon_zurich, lat_zurich) < 15

stop_business_time = stop_time.where((stop_time.arrival_time <= "20:00:00") & (stop_time.departure_time >= "07:00:00"))
print("DataFrame representing filtered stop times has " + str(stop_business_time.count()) + " rows")

stop_zurich_unwrangled = stop.filter(in_zurich(stop.stop_lat, stop.stop_lon) == True)
stop_zurich = stop_zurich_unwrangled.withColumn('parent_station_2', F.substring(stop_zurich_unwrangled.stop_id, 0, 7))
stop_zurich = stop_zurich.withColumn('stop_id_2', F.when(F.length(F.col("stop_id")) > 7, stop_zurich.parent_station_2).otherwise(stop_zurich.stop_id))
stop_zurich = stop_zurich.drop("stop_id").withColumnRenamed("stop_id_2", "stop_id")

print("DataFrame representing filtered stop has " + str(stop_zurich.count()) + " rows")

stop_time_zurich = stop_business_time.join(stop_zurich_unwrangled, on="stop_id")
stop_time_zurich_cleaned = stop_time_zurich.withColumn('parent_station_2', F.substring(stop_time_zurich.stop_id, 0, 7))
stop_time_zurich_cleaned = stop_time_zurich_cleaned.withColumn('stop_id_2', F.when(F.length(F.col("stop_id")) > 7, stop_time_zurich_cleaned.parent_station_2).otherwise(stop_time_zurich_cleaned.stop_id))
stop_time_zurich_cleaned = stop_time_zurich_cleaned.drop("stop_id").withColumnRenamed("stop_id_2", "stop_id")
print("DataFrame representing stop time in zurich in normal hours has " + str(stop_time_zurich_cleaned.count()) + " rows")

trip_in_zurich = trip.join(stop_time_zurich_cleaned, on="trip_id")
print("DataFrame representing trip in zurich in normal hours has " + str(trip_in_zurich.count()) + " rows")

calendar_business_day = calendar.filter((calendar.monday == 1) &
                            (calendar.tuesday == 1) &
                            (calendar.wednesday == 1) &
                            (calendar.thursday == 1) &
                            (calendar.friday == 1))

joined_data = trip_in_zurich.join(calendar_business_day, on="service_id")
print("DataFrame representing stop in zurich during normal hours on week days has " + str(joined_data.count()) + "rows ")

joined_data = joined_data.join(route, on="route_id")

joined_data_filter = joined_data.select(["trip_id", "stop_id", "stop_name", "arrival_time", "departure_time", "stop_sequence", "parent_station", "direction_id", "route_desc"])
joined_data_filter.show(5)

## Switching from stops to journey

We currently have a data formate where a row represents a stop.

We aim at switching to a data format where a row represent a journey of a public transport from one row to the next ones.

These rows will be the edges of our graph.

We will switch to this data format using pyspark windows.

joined_data_filter = joined_data_filter.withColumn("stop_sequence", joined_data_filter.stop_sequence.cast('int'))
trip_window = Window.partitionBy('trip_id').orderBy('stop_sequence')
stop_id_from = F.lag('stop_id', 1).over(trip_window).alias('stop_id_from')
departure_time = F.lag('departure_time', 1).over(trip_window).alias('departure_time')
stop_name_from = F.lag('stop_name', 1).over(trip_window).alias('stop_name_from')

journey_df = joined_data_filter.select("trip_id", stop_id_from, "stop_id", stop_name_from, "stop_name", departure_time, "arrival_time", "route_desc")
journey_df = journey_df.withColumnRenamed("stop_id", "stop_id_to").withColumnRenamed("stop_name", "stop_name_to")
journey_df.show(5)

@F.udf
def compute_duration(departure_time, arrival_time):
    if (departure_time is None):
        return 0
    dep_time = datetime.datetime.strptime(departure_time, "%H:%M:%S")
    arr_time = datetime.datetime.strptime(arrival_time, "%H:%M:%S")
    return (arr_time - dep_time).seconds / 60

journey_df = journey_df.withColumn("duration", compute_duration(journey_df.departure_time, journey_df.arrival_time))
journey_df = journey_df.filter(journey_df.departure_time.isNotNull())

print("We have " + str(journey_df.count())+ " possible journeys through public transportations")
journey_df.show(5)

## Walking distance

Every users can walk to a distance of 500 meters from one stop to another.

The duration of a transfer is 2 mn, on which we add 1 mn every 50 meters.

As we need to have these transfer to compute correctly our graph, we isolate it from the stops by using a cross join and computing each distance, then filtering it.

print("We have " +  str(stop_zurich.count())+ " stops")

stop_zurich_renamed = stop_zurich.select(stop_zurich.stop_id.alias("stop_id_2"), stop_zurich.stop_name.alias("stop_name_2"), stop_zurich.stop_lat.alias("stop_lat_2"), stop_zurich.stop_lon.alias("stop_lon_2"))
cross_join = stop_zurich.crossJoin(stop_zurich_renamed)
cross_join = cross_join.withColumn("stop_lat", cross_join.stop_lat.cast('float'))\
                       .withColumn("stop_lat_2", cross_join.stop_lat_2.cast('float'))\
                       .withColumn("stop_lon", cross_join.stop_lon.cast('float'))\
                       .withColumn("stop_lon_2", cross_join.stop_lon_2.cast('float'))

cross_join.show(5)

@F.udf
def haversine_distance_udf(lon1, lat1, lon2, lat2):
    """
    Calculate the great circle distance between two points 
    on the earth (specified in decimal degrees)
    """
    # convert decimal degrees to radians 
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    # haversine formula 
    dlon = lon2 - lon1 
    dlat = lat2 - lat1 
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * asin(sqrt(a)) 
    r = 6371 # Radius of earth in kilometers. Use 3956 for miles
    return c * r

@F.udf
def transfer_time(distance):
    return (2 + distance * 20)

@F.udf 
def transfer_time_minute(transfer_time):
    return floor(transfer_time / 60)

@F.udf 
def transfer_time_second(transfer_time):
    return floor(transfer_time % 60)


cross_join = cross_join.withColumn("distance", haversine_distance_udf(cross_join.stop_lon, cross_join.stop_lat, cross_join.stop_lon_2, cross_join.stop_lat_2))
cross_join = cross_join.filter(cross_join.distance < 0.5)
cross_join = cross_join.withColumn("transfer_time", transfer_time(cross_join.distance))
cross_join.show(10)

print("we have " + str(cross_join.count()) + " possible transfer by walking distance")

# Selecting transfers, converting transfer time in unix for next arrival time 
spark.conf.set("spark.sql.session.timeZone", "GMT")

stop_transfers = cross_join.select(cross_join.stop_id,cross_join.stop_name,cross_join.stop_id_2,cross_join.stop_name_2,cross_join.transfer_time)\
                            .withColumn("time_transfer_unix",F.unix_timestamp(F.col("transfer_time"),"mm"))

stop_transfers.show(5)

## Creating our exhaustive graph edges

As we want to compute our *Graph edges 2* which contains the departure time for all walking distances, we need to left outer join our public transportation DataFrame arrival id with our stop transfers.

# Selecting all stops with arrival times for left outer join 
# spark.conf.set("spark.sql.session.timeZone", "UTC")

stop_times = journey_df.select(["stop_id_to", "stop_name_to", "arrival_time"])\
                       .withColumn("departure_time_unix",F.unix_timestamp(F.col("arrival_time"),"H:mm:ss") )\
                       .withColumnRenamed("arrival_time","departure_time")

stop_times.show(5)

# Inferring arrival time based on travel time 
left_join = stop_times.join(stop_transfers, stop_times.stop_id_to == stop_transfers.stop_id, how='left')
stop_transfer_joined = left_join.withColumn("arrival_time",
                                            F.from_unixtime(F.col("departure_time_unix")+F.col("time_transfer_unix")\
                                                                            ,"HH:mm:ss"))

stop_transfer_joined = stop_transfer_joined.select(["stop_id", "stop_name", "departure_time", "stop_id_2", "stop_name_2", "arrival_time", "transfer_time"])\
                                           .withColumnRenamed("transfer_time","duration")

stop_transfer_joined.show(5)

print("we have " + str(stop_transfer_joined.count()) + " possible transfer by walking distance")

# Graph DataFrames creation

Now that we have all of our dataframes ready, we need to convert them to the correct data format for our graph. 

Thus, we need to combine the public transportation dataframe with either the walking distance one, either the stop_transfer_join previously created to get our 2 desired final dataframes representing edges.

graph_nodes = stop_zurich.select(["stop_id", "stop_name", "stop_lat", "stop_lon"])\
                   .withColumnRenamed("stop_id", "id")\
                   .withColumnRenamed("stop_lat", "lat")\
                   .withColumnRenamed("stop_lon", "lon")
graph_nodes.show(5)


graph_journey_edges = journey_df.select(["trip_id", "stop_id_from", "stop_id_to", "departure_time", "arrival_time", "duration", "route_desc"])\
                        .withColumnRenamed("stop_id_from", "src")\
                        .withColumnRenamed("stop_id_to", "dst")\
                        .withColumnRenamed("route_desc", "type")

graph_journey_edges.show(5)

graph_transfer_edges = cross_join.withColumn("trip_id", F.lit(-1))\
                                 .select(["trip_id", "stop_id", "stop_id_2", "transfer_time"])\
                                 .withColumnRenamed("stop_id", "src")\
                                 .withColumnRenamed("stop_id_2", "dst")\
                                 .withColumnRenamed("transfer_time", "duration")\
                                 .withColumn("departure_time", F.lit(None).cast(StringType()))\
                                 .withColumn("arrival_time", F.lit(None).cast(StringType()))\
                                 .withColumn("type", F.lit("transfer"))\
                                 

graph_transfer_edges.show(5)

print("graph_journey_edges has " + str(graph_journey_edges.count()) + " rows")
print("graph_transfer_edges has " + str(graph_transfer_edges.count()) + " rows")

graph_edges_1 = graph_journey_edges.union(graph_transfer_edges)

print("graph_edges has " + str(graph_edges_1.count()) + " rows")

graph_edges_1.show(5)

graph_transfer_joined_edges = stop_transfer_joined.withColumn("trip_id", F.lit(-1))\
                                                  .select(["trip_id", "stop_id", "stop_id_2","departure_time", "arrival_time", "duration"])\
                                                  .withColumnRenamed("stop_id", "src")\
                                                  .withColumnRenamed("stop_id_2", "dst")\
                                                  .withColumn("type", F.lit("transfer"))\

graph_transfer_joined_edges.show(5)

print("graph_journey_edges has " + str(graph_journey_edges.count()) + " rows")
print("graph_transfer_joined_edges has " + str(graph_transfer_joined_edges.count()) + " rows")

graph_edges_2 = graph_journey_edges.union(graph_transfer_joined_edges)

print("graph_edges has " + str(graph_edges_2.count()) + " rows")

graph_edges_2.show(5)

## Exporting DataFrames to HDFS

As we want to process our graph with networkx, we need to modelize it locally in a python kernel.

Thus, we send our graph DataFrames to the HDFS.

The code is commented as the edges were already send

graph_nodes.write.parquet('/user/vyuan/final_6/graph_nodes.parquet')

graph_edges_1.write.parquet('/user/vyuan/final_6/graph_edges_1.parquet')

graph_edges_2.write.parquet('/user/vyuan/final_6/graph_edges_2.parquet')
